LibGL+LibSoftGPU: Make DeviceInfo generic and specialize on the device
Problem:
- DeviceInfo is only implemented in LibSoftGPU, but it is used in
  LibGL. LibGL should not have a dependency on LibSoftGPU if it is
  ever going to be a generic GL engine. This tight coupling makes it
  difficult to do some metaprogramming/constexpr-programming to remove
  some run-time computations.

Solution:
- Decouple `DeviceInfo` from `Device` and move into LibGL and make an
  instance of it in LibSoftGPU which can be referenced.
- The specific instance of DeviceInfo is coupled to the
  `SoftGPU::Device` through namespace rather than included as a member
  variable of the instance. Since it doesn't change there is no need
  to be part of the instance.

Note: This is a step towards making it generic, but this does not
really remove the dependency yet.
